Added a database initialisation command to the menu
If this is our first run on a new system, it isn't
fun to have to populate the db, hunting down the
variable names etc. Now --populate will do this,
if the database doesn't already exist and if the latest
db backup exists in the db folder. Also reminds that
we need the document root set in mamp to be up and runnning
from cold